+++ CHANGELOG.md
@@ -7,6 +7,13 @@
   (this required for proper docker support when non-local ip must be specified)
 * mkdocs.resolveDocPath() never return null:
   for single version docs '.' is returned now instead of null (#41) 
+* Update packages:
+  - mkdocs 1.3.0 -> 1.4.1
+  - mkdocs-material 8.3.6 -> 8.5.7
+  - pygments 2.12.0 -> 2.13.0 
+  - pymdown-extensions 9.4 ->  9.7
+
+WARNING: minimum recommended python is 3.8
 
 ### 2.4.0 (2022-06-17)
 * Fix variables support for gradle 7.4 (#34)